AI-PoC/slack-summary-using-llama-stack/summary_llm_agent/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.2
gunicorn==21.2.0
requests-oauthlib==1.3.1 